MakeDataFile/requirements.txt
et-xmlfile==1.1.0
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
